Improve request tests.
Refs #34

Now provides an echo server on port 9001 that proxies most
calls to the intern server, except for requests to /__echo/.

Calls to /__echo/ return a JSON response that contains the
parsed query string, headers, and payload from the request.

If a responseType query parameter is specified, a response
in the format requested will be returned if available.
Currently, 'html', 'document', 'xml', 'json', and 'gif'
are acceptable values for responseType. If any format other
than JSON is requested the response body will not contain
the payload, query or headers from the request.

The X-Requested-With header is no longer being used for
testing custom headers because IE 11 was modifying the
value before sending to the server.

Added tests for different content types and to improve
coverage.