Remove now-redundant object.assign, improve lang tests
This is now covered by lang.assign.  object.is will likely be made
redundant by lang.isIdentical in the future as well.

This commit ports one of the object.assign tests over to the lang suite,
replaces some straggler assert.equal calls with assert.strictEqual, adds
some more assertions and assertion comments to the assign/mixin tests,
and adds an assertion to the partial tests to bring lang coverage to 100%.